finalyze the project

+++ frontend/src/hooks/useAdmin.ts
@@ -15,36 +15,52 @@ import type { TSubscriber } from "../types/TSubscriber";
 
 const useAdmin = () => {
   const token = useSelector((state: RootState) => state.auth.user?.token);
+  const role = useSelector((state: RootState) => state.auth.user?.user.role);
   const [zones, setZones] = useState<Zone[]>([]);
   const [subs, setSubs] = useState<TSubscriber[]>([]);
+  const [zLoading, setZLoading] = useState(false);
+  const [subLoading, setSubLoading] = useState(false);
+  const [toglleLoading, setToglleLoading] = useState(false);
 
   const dispatch = useDispatch<AppDispatch>();
 
   useEffect(() => {
-    return () => {
+    if (role && role != "admin") {
       dispatch(removeUser());
-    };
-  }, [dispatch]);
+    }
+  }, [role, dispatch]);
 
   useEffect(() => {
     if (!token) {
       return;
     }
-    const zones = getParkingState().then((data) => {
-      setZones(data);
-    });
-    const subscriptions = getAllSubscriptions(token).then((data) => {
-      setSubs(data);
-    });
+    setZLoading(true);
+    const zones = getParkingState()
+      .then((data) => {
+        setZones(data);
+      })
+      .finally(() => {
+        setZLoading(false);
+      });
+    setSubLoading(true);
+    const subscriptions = getAllSubscriptions(token)
+      .then((data) => {
+        setSubs(data);
+      })
+      .finally(() => {
+        setSubLoading(false);
+      });
   }, [token]);
 
   const handleToggleZone = async (zoneId: string, open: boolean) => {
     if (!token) return;
-
+    setToglleLoading(true);
     try {
       const res = await toggleZoneOpen(zoneId, !open, token);
       const zonesRes = await getParkingState();
+
       setZones(zonesRes);
+      setToglleLoading(false);
     } catch (error) {
       console.log(error);
     }
@@ -65,12 +81,15 @@ const useAdmin = () => {
     const rateSpecial = Number(
       (form.elements.namedItem("rateSpecial") as HTMLInputElement).value
     );
-
+    if (!rateNormal || !rateSpecial) {
+      alert("fill the value(s) please");
+      return;
+    }
     await updateCategoryRates(categoryId, { rateNormal, rateSpecial }, token);
     alert("Category rates updated!");
   };
 
-  // ✅ Add rush hour
+  //  Add rush hour
   const handleAddRushHour = async (e: React.FormEvent<HTMLFormElement>) => {
     e.preventDefault();
     if (!token) return;
@@ -79,11 +98,15 @@ const useAdmin = () => {
     const start = (form.elements.namedItem("start") as HTMLInputElement).value;
     const end = (form.elements.namedItem("end") as HTMLInputElement).value;
 
+    if (!start || !end) {
+      alert("fill the time");
+      return;
+    }
     await addRushHour({ start, end }, token);
     alert("Rush hour added!");
   };
 
-  // ✅ Add vacation
+  //  Add vacation
   const handleAddVacation = async (e: React.FormEvent<HTMLFormElement>) => {
     e.preventDefault();
 
@@ -92,6 +115,11 @@ const useAdmin = () => {
     const form = e.currentTarget;
     const start = (form.elements.namedItem("start") as HTMLInputElement).value;
     const end = (form.elements.namedItem("end") as HTMLInputElement).value;
+
+    if (!start || !end) {
+      alert("fill the time");
+      return;
+    }
 
     await addVacation({ start, end }, token);
     alert("Vacation added!");
@@ -107,6 +135,9 @@ const useAdmin = () => {
     handleToggleZone,
     handleUpdateRates,
     setCategoryId,
+    zLoading,
+    subLoading,
+    toglleLoading,
   };
 };
